requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
openai-whisper>=20230314
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3
pandas>=2.0.0
tqdm>=4.65.0
joblib>=1.3.0
soundfile>=0.12.0
